Update database structure
Removed cuisine table, so no more cuisine preferences for users. Added user_login table. There will be stored info for login and reference to the app_user table. The password will be stored hashed in sha256. When the password data is inserted in the table, the function ENCODE(DIGEST('SomePassword','sha256')) will be used.